Update README.md
Airbnb Website Clone is a simplified version of the Airbnb platform that allows users to search for, book, and manage accommodations. The project is built using the MERN stack (MongoDB, Express, React, Node.js). It includes essential features like user authentication, listing creation, editing, and deletion for hosts, along with booking functionality for guests. The website is designed to be user-friendly and responsive, providing a seamless experience across different devices